+++ src/test/java/com/codeit/mission/deokhugam/dashboard/popularreviews/service/PopularReviewAggregateServiceTest.java
@@ -1,6 +1,5 @@
 package com.codeit.mission.deokhugam.dashboard.popularreviews.service;
 
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
